requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
mlflow==2.17.0
prometheus-client==0.21.0
kafka-python==2.0.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
